requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
flake8==4.0.1
pytest==8.3.2
requests==2.25.1
selenium==4.24.0
webdriver-manager==4.0.2